feat: Added storage warnings + Fixes
git-tfs-id: [https://dev.azure.com/code-a-software]$/home/Home;C785

+++ Home/Controls/DeviceFileExplorer.xaml.cs
@@ -246,7 +246,7 @@ namespace Home.Controls
 
                 if (selectedItem is RemoteFile rf)
                 {
-                    if (ByteUnit.FromB(rf.Length) <= ByteUnit.FromMB(10))
+                    if (ByteUnit.FromB((ulong)rf.Length) <= ByteUnit.FromMB(10))
                     {
                         string ext = System.IO.Path.GetExtension(rf.Path).ToLower();
